Bang Bang Loops Again, Climber is Fixed
Basically, the wheel was in the wrong place, so we have less crush now.
Yay. Now we need to actually make the bang-bang loops work again instead
of PID. If this doesn't work then we REALLY know that it won't work. The
climber apparently had a debug button on it that Jason forgot to delete,
so that's gone now.